+++ src/utils/loadAndBundleSpec.ts
@@ -5,9 +5,9 @@ import { OpenAPISpec } from '../types';
 
 export async function loadAndBundleSpec(specUrlOrObject: object | string): Promise<OpenAPISpec> {
   const parser = new JsonSchemaRefParser();
-  const spec = await parser.bundle(specUrlOrObject, {
+  const spec = (await parser.bundle(specUrlOrObject, {
     resolve: { http: { withCredentials: false } },
-  } as object);
+  } as object)) as any;
 
   if (spec.swagger !== undefined) {
     return convertSwagger2OpenAPI(spec);
@@ -19,7 +19,7 @@ export async function loadAndBundleSpec(specUrlOrObject: object | string): Promi
 export function convertSwagger2OpenAPI(spec: any): Promise<OpenAPISpec> {
   console.warn('[ReDoc Compatibility mode]: Converting OpenAPI 2.0 to OpenAPI 3.0');
   return new Promise<OpenAPISpec>((resolve, reject) =>
-    convertObj(spec, { patch: true, warnOnly: true }, (err, res) => {
+    convertObj(spec, { patch: true, warnOnly: true, text: '{}' }, (err, res) => {
       // TODO: log any warnings
       if (err) {
         return reject(err);